+++ cadhy/blender/operators/op_build_channel.py
@@ -57,6 +57,8 @@ class CADHY_OT_BuildChannel(Operator):
                     "TRAP": SectionType.TRAPEZOIDAL,
                     "RECT": SectionType.RECTANGULAR,
                     "CIRC": SectionType.CIRCULAR,
+                    "TRI": SectionType.TRIANGULAR,
+                    "PIPE": SectionType.PIPE,
                 }
 
                 params = ChannelParams(